Refactore out of op area warning dialog implementation
