+++ opensrp-reveal/src/main/java/org/smartregister/reveal/widget/GeoWidgetFactory.java
@@ -2,6 +2,7 @@ package org.smartregister.reveal.widget;
 
 import android.app.Activity;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -34,6 +35,7 @@ import com.vijay.jsonwizard.interfaces.CommonListener;
 import com.vijay.jsonwizard.interfaces.FormWidgetFactory;
 import com.vijay.jsonwizard.interfaces.JsonApi;
 import com.vijay.jsonwizard.interfaces.LifeCycleListener;
+import com.vijay.jsonwizard.presenters.JsonFormFragmentPresenter;
 import com.vijay.jsonwizard.utils.ValidationStatus;
 import com.vijay.jsonwizard.views.JsonFormFragmentView;
 
@@ -43,6 +45,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.smartregister.reveal.BuildConfig;
 import org.smartregister.reveal.R;
+import org.smartregister.reveal.util.AlertDialogUtils;
 import org.smartregister.reveal.util.Constants.Map;
 import org.smartregister.reveal.util.RevealMapHelper;
 import org.smartregister.reveal.validators.MinZoomValidator;
@@ -58,6 +61,9 @@ import io.ona.kujaku.callbacks.OnLocationComponentInitializedCallback;
 import io.ona.kujaku.layers.BoundaryLayer;
 import timber.log.Timber;
 
+import static android.content.DialogInterface.BUTTON_NEGATIVE;
+import static android.content.DialogInterface.BUTTON_NEUTRAL;
+import static android.content.DialogInterface.BUTTON_POSITIVE;
 import static org.smartregister.reveal.util.Constants.DIGITAL_GLOBE_CONNECT_ID;
 import static org.smartregister.reveal.util.Constants.JsonForm.OPERATIONAL_AREA_TAG;
 import static org.smartregister.reveal.util.Constants.JsonForm.STRUCTURES_TAG;
@@ -81,7 +87,7 @@ public class GeoWidgetFactory implements FormWidgetFactory, LifeCycleListener, O
 
     private static com.mapbox.geojson.Feature operationalArea = null;
 
-    public static ValidationStatus validate(JsonFormFragmentView formFragmentView, RevealMapView mapView) {
+    public static ValidationStatus validate(JsonFormFragmentView formFragmentView, RevealMapView mapView, JsonFormFragmentPresenter presenter) {
 
         if (!Utils.isEmptyCollection(mapView.getValidators())) {
             for (METValidator validator : mapView.getValidators()) {
@@ -93,8 +99,31 @@ public class GeoWidgetFactory implements FormWidgetFactory, LifeCycleListener, O
                     }
                 } else if (validator instanceof WithinOperationAreaValidator) {
                     // perform within op area validation
-                    if (!validator.isValid("", true)) {
-                        Toast.makeText(formFragmentView.getContext(), validator.getErrorMessage(), Toast.LENGTH_LONG).show();
+                    if ( !validator.isValid("", true)) {
+                        if (org.smartregister.reveal.util.Utils.displayAddStructureOutOfBoundaryWarningDialog()) {
+                            AlertDialogUtils.displayNotificationWithCallback(formFragmentView.getContext(), R.string.register_outside_boundary_title, R.string.register_outside_boundary_warning, R.string.register, R.string.cancel, new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+                                    switch (which) {
+                                        case BUTTON_NEGATIVE:
+                                            dialog.dismiss();
+                                            break;
+                                        case BUTTON_NEUTRAL:
+                                            dialog.dismiss();
+                                            break;
+                                        case BUTTON_POSITIVE:
+                                          ((WithinOperationAreaValidator) validator).setDisabled(true);
+                                            presenter.validateAndWriteValues();
+                                            dialog.dismiss();
+                                            break;
+                                        default:
+                                            break;
+                                    }
+                                }
+                            });
+
+                        }
+
                         return new ValidationStatus(false, validator.getErrorMessage(), formFragmentView, mapView);
                     }
                 }
